Add new eslint rule + refactor
