+++ packages/back-end/src/routers/tag/tag.controller.ts
@@ -8,7 +8,7 @@ import { addTag, removeTag } from "back-end/src/models/TagModel";
 import { removeTagInMetrics } from "back-end/src/models/MetricModel";
 import { removeTagInFeature } from "back-end/src/models/FeatureModel";
 import { removeTagFromSlackIntegration } from "back-end/src/models/SlackIntegrationModel";
-import { removeTagInAttribute } from "back-end/src/routers/attributes/attributes.controller";
+import { removeTagInAttribute } from "back-end/src/services/attributes";
 import { removeTagFromExperiments } from "back-end/src/models/ExperimentModel";
 
 // region POST /tag
